Clean up 035 example test support
