+++ examples/035-mdm-tux-rs/src/bin/target.rs
@@ -2078,56 +2078,29 @@ fn parse_provider_override(args: &[String]) -> anyhow::Result<Option<ProviderKin
 
 #[cfg(test)]
 mod tests {
+    mod test_support {
+        include!(concat!(env!("CARGO_MANIFEST_DIR"), "/src/test_support.rs"));
+    }
+
     use super::{
         build_target_runtime_surface, discard_live_session_with_mob_cleanup,
         parse_provider_override,
     };
-    use async_trait::async_trait;
     use mdm_tux::ProviderKind;
     use meerkat::{
         AgentFactory, FactoryAgentBuilder, LlmClient, SessionAgentBuilder, SessionService,
         encode_llm_client_override_for_service,
     };
-    use meerkat_client::types::LlmStream;
-    use meerkat_client::{LlmError, LlmRequest, TestClient};
+    use meerkat_client::TestClient;
     use meerkat_comms::agent::CommsToolDispatcher;
     use meerkat_comms::identity::Keypair;
     use meerkat_core::Config;
     use meerkat_core::service::{CreateSessionRequest, InitialTurnPolicy, SessionBuildOptions};
     use meerkat_core::types::ContentInput;
     use meerkat_mob_mcp::{AgentMobToolSurfaceFactory, MobMcpState};
-    use std::sync::{Arc, Mutex};
+    use std::sync::Arc;
+    use test_support::CaptureClient;
     use tokio::sync::mpsc;
-
-    #[derive(Default)]
-    struct CaptureClient {
-        inner: TestClient,
-        seen_tools: Mutex<Vec<String>>,
-    }
-
-    impl CaptureClient {
-        fn tool_names(&self) -> Vec<String> {
-            self.seen_tools.lock().expect("capture lock").clone()
-        }
-    }
-
-    #[cfg_attr(target_arch = "wasm32", async_trait(?Send))]
-    #[cfg_attr(not(target_arch = "wasm32"), async_trait)]
-    impl LlmClient for CaptureClient {
-        fn stream<'a>(&'a self, request: &'a LlmRequest) -> LlmStream<'a> {
-            *self.seen_tools.lock().expect("capture lock") =
-                request.tools.iter().map(|tool| tool.name.clone()).collect();
-            self.inner.stream(request)
-        }
-
-        fn provider(&self) -> &'static str {
-            self.inner.provider()
-        }
-
-        async fn health_check(&self) -> Result<(), LlmError> {
-            self.inner.health_check().await
-        }
-    }
 
     #[test]
     fn provider_override_requires_model_and_provider_together() {
